Length of wiki_references referenced_name
Another minor database migration. The length of the
'referenced_name' field in the previous schema was
limited to 60 characters. Increase that to 255
characters (matching the length of then 'name' field
in 'pages', etc).

This has no effect in SQLite3 users (the default), since
SQLite3 does not enforce these length restrictions. But
MySQL users need this.

As always, run

   rake upgrade_instiki

to seamlessly upgrade your database to the latest schema.