Revamp support for HTML5 <video> and <audio> Elements
Instead of a single src attribute on the
<video> or <audio> element,
allow for multiple <source> child-elements.

The revised syntax allows for a whitespace-separated
list of media files to upload.

Unfortunately, there's no provision for setting the type
attribute (e.g., inferring it from the file extension).
Which, apparently, means that the client downloads each
version of the video until it finds one it can support.