Fix load path and filename problem.
A missing "#" in the application.rb autoload paths meant that the lib/chunks
code wasn't working, but this didn't seem to behave when "wiki.rb" defined
a top-level "WikiChunks" module (likewise "uri.rb"). Renamed to fix. A few
other patches made to cope with this and address other arising issues.