Don't double-render
When the user saves a page (either a new one or a revision of an old one),
it gets rendered TWICE. The first time, is in the revise action (where we
want to update the WikiReferences, so we need to build the page in order
to figure out what those ARE). Then, the user gets redirected to the "show"
action, where the page gets rendered AGAIN.

That's inefficient.

Instead, we cache a copy of the rendered content and
retrieve it in the "show" action (and then delete the cached copy). That
way, each "save" only results in the page being rendered once.